Modified the maven file to automatically generate the Ragel file, run tests on the generated Java file, and reports those.
Currently, there are some test failures. We should agree if they are true
positives (where positive means a test failure).

There are also some refactoring I'd be willing to do to the ragel file,
after reading on URI schemes, but we need to discuss that (if necessary --
can I have dictatorship on this repository?).