Added unit-tests for regex equivalents of the UrlParserTest.
A few tests were slightly modified, or removed, to account the different
usability between the Url and Matcher objects. There are also a few things
that the regex seems to do better, while the Url object do others better.
Overall, the UrlParser support a wider range of grammar.

The tests as they are though are not necessarily representative of what we
want to support. Some failures should be failures, as they are more of URI
that differ so much from a URL that we might now want to support that
anyway. On the other hands, there might be variations that we want to
support for pragmatic purpose. Maybe we should open issues for each
failure and vote if it makes sense to support it.